python/requirements.txt
python-dotenv==1.0.1
hamming-sdk==0.0.12
openai==1.14.2
